feat: Refactor order creation and add stop order functionality
- Add support for creating stop orders
- Modify several files to include new functionality
- Refactor `WorkerOrder` struct and functions to support Decimal and Currency types
- Add methods to `Database` trait for reading trading vehicles and creating stop orders for them
- Add validation to quantity field in trade dialog prompt

+++ trust-db-sqlite/src/schema.rs
@@ -84,7 +84,7 @@ diesel::table! {
         updated_at -> Timestamp,
         deleted_at -> Nullable<Timestamp>,
         price_id -> Text,
-        quantity -> Integer,
+        quantity -> BigInt,
         trading_vehicle_id -> Text,
         action -> Text,
         category -> Text,
